# project/static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Notes API",
    "version": "1.0.0",
    "description": "API for managing notes"
  },
  "paths": {
    "/api/notes": {
      "get": {
        "summary": "List all notes",
        "responses": {
          "200": {
            "description": "A list of notes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Note" }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new note",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/NoteInput" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Note created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Note" }
              }
            }
          }
        }
      }
    },
    "/api/notes/{note_id}": {
      "get": {
        "summary": "Get a note by ID",
        "parameters": [
          {
            "name": "note_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "The note",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Note" }
              }
            }
          },
          "404": {
            "description": "Note not found"
          }
        }
      },
      "put": {
        "summary": "Update a note",
        "parameters": [
          {
            "name": "note_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/NoteInput" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Note updated"
          },
          "404": {
            "description": "Note not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a note",
        "parameters": [
          {
            "name": "note_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Note deleted"
          },
          "404": {
            "description": "Note not found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Note": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "content": { "type": "string" }
        }
      },
      "NoteInput": {
        "type": "object",
        "properties": {
          "content": { "type": "string" }
        },
        "required": ["content"]
      }
    }
  }
}